--- a/ChronoGov/.gitattributes
+++ b/ChronoGov/.gitattributes
@@ -1,0 +1,3 @@
+tests/** linguist-vendored
+vitest.config.js linguist-vendored
+* text=lf