--- a/ChronoGov/Clarinet.toml
+++ b/ChronoGov/Clarinet.toml
@@ -1,0 +1,19 @@
+[project]
+name = 'ChronoGov'
+description = ''
+authors = []
+telemetry = true
+cache_dir = '.\.cache'
+requirements = []
+[contracts.ChronoGov]
+path = 'contracts/ChronoGov.clar'
+clarity_version = 3
+epoch = 3.1
+[repl.analysis]
+passes = ['check_checker']
+
+[repl.analysis.check_checker]
+strict = false
+trusted_sender = false
+trusted_caller = false
+callee_filter = false