--- a/ChronoGov/contracts/ChronoGov.clar
+++ b/ChronoGov/contracts/ChronoGov.clar
@@ -1,0 +1,30 @@
+
+;; title: ChronoGov
+;; version:
+;; summary:
+;; description:
+
+;; traits
+;;
+
+;; token definitions
+;;
+
+;; constants
+;;
+
+;; data vars
+;;
+
+;; data maps
+;;
+
+;; public functions
+;;
+
+;; read only functions
+;;
+
+;; private functions
+;;
+